+++ spec/workers/property_synchronisation_spec.rb
@@ -277,7 +277,7 @@ RSpec.describe Workers::PropertySynchronisation do
       expect(sync.stats[:properties_deleted]).to eq 0
     end
 
-    it "registers updates and deletions when successful" do
+    it "registers updates, skips and deletions when successful" do
       stub_call(:delete, "https://api.roomorama.com/v1.0/host/disable")    { [200, {}, [""]] }
 
       create_property(host_id: host.id, identifier: "prop1", data: roomorama_property.to_h)
@@ -300,6 +300,9 @@ RSpec.describe Workers::PropertySynchronisation do
         Result.new(roomorama_property)
       }
 
+      # create
+      subject.skip_property
+
       # prop3 is not included - should be deleted
       #
       expect {
@@ -315,6 +318,7 @@ RSpec.describe Workers::PropertySynchronisation do
       expect(sync.stats[:properties_created]).to eq 1
       expect(sync.stats[:properties_updated]).to eq 1
       expect(sync.stats[:properties_deleted]).to eq 1
+      expect(sync.stats[:properties_skipped]).to eq 1
     end